Complex numbers are plugged into numeric type classes.
There are some things that need to be reviewed, e.g. the definition
I chose for complex quotient/modulus. Also, we *really* need to
start adding some extra capabilities to Fractional so that complex
numbers can stop flying on toDouble/fromDouble in order to work.

That said, I'm pretty pleased with how this seems so far. I also
threw in Semigroup and Monoid because they were easy and (might)
appease the fp constituency a bit.