Add limiting methods (eg. limitToInt) to Rational.
This adds a new set of methods to limit Rationals to a smaller form.
This includes limitTo(BigInt) to limit both the numerator and
denominator to be below an arbitrary magnitude,
limitDenominatorTo(BigInt), which only limits the denominator, and
limitToLong and limitToInt for special treatment of these cases.

This also adds tests for the new functionality.