Shore up type class definitions and members.
There are two somewhat conflated purposes in this commit.
The first is to add support for Byte and Short to Eq,
Order and Ring. The second is to try to reduce the
number of methods that a typeclass instance must implement.

After doing some testing to confirm that overriding
methods in traits doesn't cost us anything, we can go
ahead and create "minimal" definitions of things like
times in Ring[A] in terms of plus.