Fix Field.isWhole, and rename eq to eqv.
Turns out there was a subtle bug in Field--we were calling eq()
even though Field[A] no longer extends Eq[A]. This wasn't caught
as a bug because AnyRef.eq was stepping in.

So, I have done s/eq/eqv/ to refactor eq,neq,lteq and gteq.
I also removed the Field.isWhole implementation (which I couldn't
see how to do without Eq) and implemented it manually in the
type class instances.