Start using Trig[A] type class in Complex[A].
This removes some situations where we were losing
precision in complex numbers, which will help when
using types more precise than Double.

The one remaining place where we convert to Double
is where we need to do A ** A (since Field only
supports A ** Int). It would be nice to be able
to add pow(A, A) to Field someday.