Get Spire basically working with 2.10.0-M3
Currently ScalaTest dependency is disabled, so I can't run the tests.
At this point all I know is that it compiles and some of the benchmarks
run successfully.

Specialization has been disabled due to SI-5788 breaking a lot of our
recursive code. Hopefully it will get fixed; if not, we'll need to
restructure some things.