Fix zero denominator errors.
It seems like we actually use "invalid" rational numbers internally in some
cases. After a bit of thinking, I opted to create a protected factory method
called buildUnsafe which can be used in these cases.

If there is a legitimate need for end users to create invalid rationals (i.e.
Rational(1, 0)) then this decision will need to be revisited.