Remove macro-related boilerplate.
These changes use the technique suggested by Jason Zaugg and Eugene Burmako
to infer what method name to use for each operator based on the Context object
provided to the macro. This is pretty elegant: we can switch on the symbol of
the method whose definition is a macro, and then decide which method of the
implicit evidence to call.

For instance:

  class FooOps[A](lhs:A)(implicit ev:Foo[A]) {
    def + = macro Ops.binop[A, A]
  }

When the binop macro is invoked, it will find that its "parent" is the $plus
method. We'll use a mapping from operator names to method names to go from
"$plus" to "plus" and then dispatch the call to the "ev.plus" method.

The only real downside is that we can't define two different Ops classes who
want to dispatch the same operator symbol to different methods. This isn't too
bad though since it would be really confusing to use different names for the
same symbol (i.e. "add" and "plus" for +).