Specialization is go!
This commit does some work to re-enable specialization. The first step was to
reimplement @tailrec methods as manual while loops... these are very slightly
less efficient in terms of bytecode but basically fine. Next, specialization
was turned on.

Finally, all the Ops classes were unspecialized, since they are now only used
during compile-time to trigger macro evaluation. This is a nice side benefit.

(There are also a few odds-and-ends in this commit; some things moved around.)