Basic radix support as macro.
This adds a few radixes (bases 2, 8, 10, and 16) as well as a generalized
macro target, spire.macros.Macros.radix, which can support any radix between
2 and 36.

The syntax is:

  import spire.math.Radix._

  x2"1000" // evaluates to 8:Int
  x8"1000" // evaluates to 512:Int
  x16"1000" // evaluates to 4096:Int

Basically, to define your own radix (i.e. 13) you just need to say:

  import spire.macros.Macros
  implicit def myradix = new { def x13 = macro Macros.radix }

The name is important... currently it expects a single character ("x") followed
by the base number to use. This might change based on usage/feedback but for
now it seemed like a good way to minimize boiler-plate.

With all this 2.10 work I'm beginning to think we need a spire.syntax package!