Add implementation of cfor macro.
Taking inspiration from the version in Macrocosm, this version is a bit more
efficient (avoiding intermediate val definitions) and should correspond exactly
to the "ideal" while-loop version.

This commit also adds a "deps" project, which will contains macros that aren't
reliant on Spire's types/methods. The hope is that macros like cfor will make
some of Spire's implementations (and benchmarks) easier to read and write.