Added EuclideanRational and first pass at Natural.
EuclideanRational is like Rational but works for any type that has
a EuclideanRing instance. Due to its generality it's a bit more
limited, but it does have its own typeclass instances.

Natural is a somewhat weird bird: it's a cross between Haskell's
Nat type (slightly more advanced than Zero | Succ, it uses a binary
encoding) and java.math.BigInteger. Instead of being base-2, it's
base-2^32: each of its digits is a UInt. Unlike BigInteger it doesn't
use an array, but rather is an immutable cons list of values.

Interestingly, when summing numbers, if your numbers are 128 bits or
less, Natural is faster than BigInt (tested for both 1K and 1M
numbers). Beyond that BigInt wins... it seems like BigInt has a lot of
fixed overhead but a much better curve as bit-count increases (APFloat
is an even more extreme example of this).

However, due to using a cons structure instead of an array, many other
operations are slower due to being O(n) for the number of digits.