Start to rationalize support for Fractional[A].
This commit fixes some ambiguous implicit problems with Fractional[A].
The Integral/Fractional/Numeric type classes haven't seen so much
attention and need a bit more love.

This commit adds support for literal Doubles when working with Fractional
(analogous to the support for literal Int with Rings). Arguably we should
move this to Field instead, but it works for now.

The reason it might make sense to move to Field is that this commit still
doesn't support using literal doubles with Numeric[A].