Remove generic Eq from default implicits.
This removes the generic Eq which makes an Eq[A] available for ANY type
A. While the implicit was low priority, it made it easy to make mistakes
like x === 0, where x is some type A. Eq[Any] would be inferred and we'd
use Scala's == rather than what we'd exepct.

The generic eq still lives, but it has been moved to
spire.optional.genericEq, which you must explicitly import to get.