Add cfor benchmarks that update an array.
This checkin adds a set of basic benchmarks that updates each element
in an array after multiplying the previous value by 2.

This benchmark was added to replicate an external spire benchmark in
which cfor seemed to be slower than while.  The benchmark results,
however, vary enough to be inconclusive at this point.  For example,
in the case where n = 1m (as in this benchmark), the 'while' case
on my machine varies between 3-4.5ms and the tailrec & cfor benchmarks
vary between 4-6 ms.