Get faster approximations for pi, e, etc.
This commit uses better algorithms for calculating pi and e. It also
introduces an approximation for exp(k). Unfortunately, there are some
small precision problems in the exp() implementation--occasionally
it gets the last digit wrong. Still needs work but much better than
the previous strategy: Doubles and log(a*b) = log(a) + log(b).

In order to "fully support" BigDecimal we still need:

  1. Fixed exp()
  2. Implement reasonable log()
  3. Implement x**y as exp(y * log(x))
  4. Implement sin, cos, tan, etc.