Upgrade to 2.10.1 and downgrade to SBT 0.12.0.
This commit updates Spire to use the newest Scala release.

Unfortunately, the combination of Scala 2.10.1 and SBT 0.12.2/3
seems to force full recompilation on almost every change. Downgrading
to SBT 0.12.0 seems to have fixed it.

The 2.10.1 pattern matcher is pickier than 2.10.0 and some cosmetic
changes were required to suppress warnings (in particular, some classes
defined in objects needed to be referenced as Object.Class to appease
the pattern matcher). There were also some deprecations in macros
which had to be updated.

(For now the branch will continue to be called 2.10.0 but obviously
it's really more of a 2.10.x branch at this point.)