Add new approximate-equality operator.
This adds a new *approximate* equality operator, `x =~= y`, that is
based on `MetricSpace`s. To use it, an implicit `MetricSpace.Epsilon` is
required to be in scope. Two points `x` and `y` are considered
approximately equal if the distance between them is less than or equal
to the value of the implicit epsilon.

This also adds explicit implicit NormedVectorSpaces for `seq` and
`array`.