Add new checked Int/Long arithmetic macro.
This is the first pass at a new macro that rewrites Int/Long arithmetic
so that it checks for overflows. Then, at runtime, if an overflow is
detected, an `ArithmeticException` is thrown. In code where overflow is
a problem, this let's people write simple arithmetic expressions with
primitive Ints or Longs and still get overflow checking.

Currenty this covers overflow checks for `+`, `-`, `*`, `/`, and
`unary_-`.