Use Checked functions vs. hard-coding overflow checks.
This switches to calling methods on the Checked object to perform the
checked arithmetic, rather than hard-coding the checks. This will be
nearly as fast and easily inline-able by the JVM, but keep bytecode
short. It also simplifies the AST transformer.

This also uses quasiquotes to simplify things as well.

We must widen the types first before performing a weak_<:<, so that is
fixed now too.

The local attrs are now reset after the rewrite too. This may be able to
be removed when SI-6711 is fixed.