Infinite set examples.
Here are two interesting Set implementations that give up the ability
to act as containers in exchange for the ability to represent infinite
sets.

MathSet is more conservative: it represents either an explicit, finite
set of elements, or an implicit universe with an explicit, finite set
of elements removed.

PureSet basically just reifies A => Boolean as a set. As such, it has
no idea which members (if any) it contains. However, this also makes
it easy to define operations like cartesian product.

Both sets have some methods that require a universe value, specifying
the list of all possible members of the set.