Random orthogonal matrices drawn from the Haar measure
As part of it, create a few classic random variables, which are just thin
wrappers around spire.random.Generator's, with a different philosophy
that I find more useful for the application I have in mind in spire.matrix.
It may make sense to try to incorporate them in spire.random in the future.

Note that the tests require the Java library JDistlib:
    http://jdistlib.sourceforge.net
A Jar may be found at
    https://bitbucket.org/alesia/jdistlib/downloads/jdistlib-0.0.7.jar