Get reciprocal working properly, add tests.
This commit does change Interval's implmentation: replacing the
division implementation (now determined to be buggy) with an
implementation using * and reciprocal. This will slightly degrate
interval bounds for "lossy" types like Double, but is much simpler
to implement correctly and test.

I am not giving up on a direct division method, but for now it's
commented out. Now that we have a well-tested reciprocal, it will
be easier to write ScalaCheck tests for division.