Change organisation of naive vs fast BLAS 3 operations
Rename “fast” to “layered” and BLAS.LayeredBLAS3.gemm does not call
BLAS.NaiveBLAS3.gemm for small matrices. The change is motivating by
testing (so that LayeredBLAS3 operations are tested for more matrix
sizes and shapes). We will later put an umbrella on the top of naive
and layered algorithm to get the best performance for all sizes.