Add basic sketch of a random monad.
Dist[A] is already a monad but requires a generator to be threaded
through it to produce values. The new Random[A, G] monad is also
parameterized on a generator, wraps a Dist[A], and has methods to
automatically thread the generator context through.

This hopefully frees us from the immutable Generator types.