Remove Ops macros in favor of Machinist.
Now that we are publishing the ops macros externally via Machinist,
it made sense to stop duplicating that code internally. Machinist
0.3.0 supports all the operator styles/names that Spire needs as
of now, and it will be easy to publish future changes as we need
them for Spire.

(We could also keep our own names map here, if we wanted to stop
relying on Machinist's DefaultOps class.)