Potential refactor to MersenneTwister32.
This refactor does a few things:

 1. Reduces the size of each Generator by removing many fields.
 2. Ensures that saved/restored RNG state works correctly.
 3. Makes constructor package private, creates named constructors.
 4. Improves some constant inlining.
 5. Change companion's "seed" type param to (Int, Array[Int]).

The issue with the previous code was that the constructor would
always reinitialize some fields, even when restoring previous
state via setSeedBytes. Also, the companion's "seed" type parameter
was changed to reflect the actual saved state in the PRNG.

Before this change is applied to the 64-bit version (and potentially
merged) I'd like to get Dusank's opinion on the changes, since it
is his code. I attempted not to change any of the actual RNG logic
(obviously) but it would be great to get the changes reviewed.