Reduce spire.math._ imports.
Since spire.math is such a big package, this change makes
it easier to be clear on what is actually being used.
Especially with methods like max (which are provided by
scala.math, spire.math, and also many types) it can be
confusing to use a wildcard import.

Hopefully this will also make recompilation a bit easier.