Bump scalaVersion to 2.11.2 in Build.scala
This commit finally gets Spire using 2.11 by default. The reason
we took so long getting here was all the warnings that our build
would spit out. This commit fixes all but 9 of the warnings.

Many warnings (unused imports, BigDecimal, etc) were dealt with
directly. However, there were a lot of deprecation warnings
involving macros where there was no obvious way to be compatible
with 2.10 but to ellide warnings in 2.11.

This commit solves the problem (well, mostly solves it) using
spire.macros.Compat, an object that contains methods implementing
the changed/deprecated methods (as well as a type alias for the
deprecated Context type). This means instead of hundreds of warnings
in the macros project, we now have 8, and we have 0 warnings
in core.

Sadly, there is still the persistent "Not a simple type" warning
as well.