Override hashCode separately in LongRational and BigRational
In LongRational, we know that the numerator is within Long range, so we
always use unifiedPrimitiveHashcode if the denominator is 1. We also have
to override longValue, since that is used by unifiedPrimitiveHashcode.

In BigRational, we know that the numerator is larger than a long if the
denominator is 1, so we never have to use unifiedPrimitiveHashcode

Fixes #368