Adds a fast path for comparisons between LongRationals
Adds a fast path that does a simple comparison without calculating a gcd
and without wrapping to SafeLong. In case of overflow, the generic code
will be executed.

This improves performance of comparisons between small to medium rationals
significantly, while having only an insignificant negative performance
impact in the overflow case.

LongAlgebra.compare is used for the comparison since java.lang.Long.compare
is not availble prior to java 7. This should be switched to java.lang.Long.compare as soon
as java 6 compatibility is dropped.

Fixes #370