Fix bugs with ULong#toFloat and ULong#toDouble.
This commit adds tests in terms of BigInt to be sure that we are
generating a reasonable Float/Double value for ULong values.

Ideally we would write some custom floating point code to avoid
allocating BigInt/BigInteger instances when calling toFloat and
toDouble. However, this code fixes the immediate problem, and
only allocates those in cases where the ULong value is too large
to use code for Longs.

ULong#toString has a related issue -- it still constructs a
BigInt value.

Fixes #387