Unify and fix Polynomial's equals and hashCode methods.
This commit adds a termsIterator method to Polynomial. This method
returns an Iterator[Term[C]] for all non-zero coefficients in the
polynomial, from smallest-exponent to largest. This iterator is
used to implement equality and hashCode in a standardized way that
works for dense and sparse polynomials.

The hashing formula used is essentially:

  val it: Iterator[Term[C]] = p.termsIterator
  val ns: Iterator[Int] = it.map(t => (0xfeed1257 * t.exp) ^ t.coeff.##)
  val hashCode: Int = ns.foldLeft(0)(_ ^ _)

One useful side-effect is that if the terms iterator is empty
(i.e. the polynomial is zero) the hash code is 0. Also, if p
consists of a single term whose exponent is zero, the overall
hashCode will be identical to the hash code of the term's
cofficient.

One gotcha with the termsIterator method is that it uses a
(_ == 0) test to filter out terms whose coefficient is zero.
This necessary in order to support universal equality, and is
documented in the method's ScalaDoc comment. We could choose
to make this method private to Spire if it is deemed to be
"too dangerous".