Establish basic blocks for new spire.stats package
In the present structure, random.Dist is a way of producing arbitrary
values from a Generator. However a probability distribution should be a
way of assign a probability to each measurable subset of the possible
outcomes of a random experiment [1], i.e. you should be able to ask it
for pdf, cdf, mean etc. The mechanism for production of values according
to a probability distribution function should in reality be a random
variable [2].
With the proposed separation it should be possible to characterize a
probability distribution and still be able to modify the production of
values from the Generator (the role of random.Dist would basically go to
stats.RandomVariable).
stats.Distribution and stats.RandomVariable would be immutable.

[1] http://en.wikipedia.org/wiki/Probability_distribution
[2] http://en.wikipedia.org/wiki/Random_variable