Enrich Generator to return based on type
Enrich Generator to return the next random value based on a specified
type. That way one can ask for a next[Double] or next[Int] instead of
nextDouble or nextInt.
However there's a name clash, therefore the method name used is
next0[A].
Should this method be included in the Generator class itself?