Rename scalacheck-binding to laws, etc.
This commit simply moves the files from one place to another.
The laws are now located in a laws module/directory, and will
be published as spire-laws.

It also removes the Arbitrary[_] instances from ArbitrarySupport
(located in the tests project) and instead moves the relevant logic
to Gen[_] instances inside of spire.laws.gen and Arbitrary[_]
instances inside of spire.laws.arb.

It updates some of the instances to be a bit more idiomatic
and removes some potential problems. It also beings the work of
creating a wide variety of flexible Gen[_] instances to remove the
need to create so many ad-hoc Arbitrary[_] instances in our tests.

Finally, it begins standardizing around the SpireTests and SpireProperties
traits. These traits disable ScalaTest's === to avoid conflicts, and
start to standardize on the use of ScalaTest's matchers (e.g. shouldBe)
over the use of assert().

(The difference between SpireTests and SpireProperties largely comes down to
whether you use test("xyz") or property("xyz"), and whether you are using
ScalaCheck to do property-based testing.)