An attempt to improve the rational arbitrary
Also some tests for the more tricky arbitraries such as rational, safeLong
and interval. The tests check that the arbitraries generate a realistic
distribution of the possible kinds of rationals, intervals or safelongs

They are made deterministic by setting the random seed, so there should not
be any additional transient failures.