Fix root refinement bug.
The problem here was that in our method where we adjust the upper and
lower bounds so that they don't lie directly on a root. We expect that
the bounds are open, so this helps us get around this weird edge case.
The problem was that we were caching our evaluation of the polynomial at
the lower/upper bounds prior to adjusting them, so they'd remain 0. This
just ensures we don't cache the value of a bound we are no longer using.