Fix various bugs in root isolation/refinement.
This fixes a few bugs:

1) A bug in the detection of roots in between an exact rational bound
and a very close approximation. Usually the root would fall in between 2
approximations, and even then, in this specific case, the root detection
usually still reported the correct result.

2) Bug in the transformation of a root upper bound to an isolated root
bound. Originally, it was using the inverse Mobius transform, but
should've been using the Mobius transform as-is.

3) A bug in the computation the upper bound of real roots of a
polynomial. This was an off-by-1 bug that affected cases where
roots fell very near the actual upper bound - within 1 bit - and
the largest coefficient was on a term of degree <= 2.

Most importantly, this re-enables the rational roots property test for
Algebraic, which was the scalacheck test that found all 3 of these bugs.